Add transforms for self-supervised learning
-A modified compose function is added which allows mixed keynames to
exist within the compose transforms even if the input dict does not
contain the all keynames specified in the compose transform. This could
be used for self-supervised and supervised learning.
-Modify CreateConstantFlag to only create a new keyname in input dict
when it initially exist in the input dict or when the user would really
like to add a new elements in the input dict.